moved caching functions to new file
